requirements.txt
python-dotenv>=1.0,<2.0
SQLAlchemy>=2.0,<3.0
python-json-logger>=2.0,<4.0
requests>=2.32,<3.0
beautifulsoup4>=4.12,<5.0
openai>=1.68,<2.0
anthropic>=0.49,<1.0
google-generativeai>=0.8,<1.0
pywin32>=308; platform_system == "Windows"
